+++ src/components/automation/AchievementUnlocker.jsx
@@ -29,8 +29,8 @@ export default function AchievementUnlocker({ activePage }) {
     useEffect(() => {
         const darkThemes = ['dark', 'midnight', 'amethyst', 'emerald', 'cherry', 'cosmic', 'mint', 'arctic', 'nightshade'];
         setImageSrc(darkThemes.includes(theme) ?
-            `https://raw.githubusercontent.com/zevnda/steam-game-idler/refs/heads/main/public/dbg.webp`
-            : `https://raw.githubusercontent.com/zevnda/steam-game-idler/refs/heads/main/public/lbg.webp`);
+            'https://raw.githubusercontent.com/zevnda/steam-game-idler/refs/heads/main/public/dbg.webp'
+            : 'https://raw.githubusercontent.com/zevnda/steam-game-idler/refs/heads/main/public/lbg.webp');
     }, [theme]);
 
     useEffect(() => {