+++ src/components/ui/GameCard.jsx
@@ -1,4 +1,5 @@
-import { Fragment, useContext } from 'react';
+import { Fragment, useContext, useEffect, useState } from 'react';
+import { useTheme } from 'next-themes';
 import Image from 'next/image';
 
 import { AppContext } from '@/src/components/layout/AppContext';
@@ -10,7 +11,14 @@ import { Button } from '@heroui/react';
 import { TbAwardFilled, TbPlayerPlayFilled } from 'react-icons/tb';
 
 export default function GameCard({ item, setSettingsModalOpen }) {
+    const { theme } = useTheme();
     const { setAppId, setAppName, showAchievements, setShowAchievements } = useContext(AppContext);
+    const [isDarkMode, setIsDarkMode] = useState(true);
+
+    useEffect(() => {
+        const darkThemes = ['dark', 'midnight', 'amethyst', 'emerald', 'cherry', 'cosmic', 'mint', 'arctic', 'nightshade'];
+        setIsDarkMode(darkThemes.includes(theme));
+    }, []);
 
     const handleImageError = (event) => {
         event.target.src = '/fallback.jpg';
@@ -28,7 +36,7 @@ export default function GameCard({ item, setSettingsModalOpen }) {
                         priority={true}
                         onError={handleImageError}
                     />
-                    <div className='absolute flex items-center justify-evenly inset-0 bg-black bg-opacity-0 dark:bg-opacity-20 group-hover:bg-opacity-40 dark:group-hover:bg-opacity-50 transition-opacity duration-200'>
+                    <div className={`absolute flex items-center justify-evenly inset-0 bg-black bg-opacity-0 group-hover:bg-opacity-40 transition-opacity duration-200 ${isDarkMode && 'bg-opacity-20 group-hover:bg-opacity-50'}`}>
                         <div className='absolute flex justify-center w-full bottom-0 left-0 px-2 pb-0.5 opacity-0 group-hover:opacity-100 duration-200'>
                             <p className='text-sm text-offwhite bg-black bg-opacity-50 rounded-sm px-1 select-none truncate'>
                                 {item.name}