+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## Changelog
 - Fixed an issue where the `blossom` theme was not being removed correctly
 - Fixed an error with `reset settings` and `clear data` functions
-- Add styling to `in list` toggle button
+- Add styling to `in list` toggle button in `edit list modal`
+- Dim `game card` images for all dark themes